feat: update project on 11/07/2022

+++ src/app/useCase/Beverages/DeleteBeverage/DeleteBeverageUseCase.ts
@@ -4,6 +4,11 @@ export class DeleteBeverageUseCase {
   constructor(private beveragesRepository: IBeverageRepository) {}
 
   async execute(id: string) {
+    const beverageExists = await this.beveragesRepository.findById(id)
+    if (!beverageExists) {
+      throw new Error('Beverage id not found')
+    }
+
     await this.beveragesRepository.delete(id)
   }
 }
